+++ packages/client/src/store/__tests__/CallState.test.ts
@@ -69,7 +69,7 @@ describe('CallState', () => {
 
       const ps2 = state.participants;
       // should resolve in initial - non-mutated state as set at the beginning
-      expect(ps2.map((p) => p.name)).toEqual(['A', 'B', 'C', 'D', 'E', 'F']);
+      expect(ps2.map((p) => p.name)).toEqual(['F', 'B', 'E', 'A', 'C', 'D']);
     });
 
     it('should support custom sorting', () => {